requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.5.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.0
pyasn1==0.4.8
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pymongo==4.12.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
